+++ internal/errors/egeneric/egeneric.go
@@ -1,4 +1,4 @@
-package errors
+package egeneric
 
 import (
 	"fmt"
